refactor(stringcalculator): 피드백 반영 리팩토링
- 계산기 의존성 주입용 설정 클래스 추가
- Number 일급 객체의 값을 실수 타입(Double)로 변경
- StringCalculator 싱글톤 -> 일반 컴포넌트로 변경
- 문자열 계산기 최종 병합 로직 수정

+++ src/test/kotlin/stringcalculator/NumberTest.kt
@@ -12,7 +12,7 @@ class NumberTest {
     @ParameterizedTest
     @ValueSource(strings = ["0", "1", "2", "10", "100", "1000", "10000"])
     fun `유효한 타입의 인자를 전달하면 정상적으로 생성된다`(input: String) {
-        val intInput = input.toInt()
+        val intInput = input.toDouble()
         val numberFromString = Number(input)
         val numberFromInt = Number(intInput)
